requirements.txt
fastapi
uvicorn
python-multipart
langchain-google-genai
langchain-community
langchain-text-splitters
faiss-cpu
pypdf
python-docx
beautifulsoup4
requests
python-dotenv
pydantic
google-generativeai
